+++ lib/ref_inspector/supervisor.ex
@@ -5,6 +5,8 @@ defmodule RefInspector.Supervisor do
 
   use Supervisor
 
+  alias RefInspector.Config
+
 
   @doc """
   Starts the supervisor.
@@ -16,6 +18,11 @@ defmodule RefInspector.Supervisor do
 
   @doc false
   def init(_default) do
+    :ok = case Config.get(:init) do
+      nil          -> :ok
+      { mod, fun } -> apply(mod, fun, [])
+    end
+
     options  = [ strategy: :one_for_one, name: __MODULE__ ]
     children = [
       RefInspector.Pool.child_spec,